RoutineAdapter moved to .routine package
